pyproject.toml
[project]
name = "wiki-rag"
description = "A simple RAG implementation using hybrid search with context optimisations against MediaWiki articles."
keywords = ["rag", "llm", "hybrid", "search", "context", "milvus", "mediawiki", "ai", "research"]

license = "BSD-3-Clause"
license-files = ["LICENSE"]
readme = "README.md"

authors = [
    {name = "Eloy Lafuente (stronk7)", email = "stronk7@stronk7.com"},
]

classifiers = [
    "Typing :: Typed",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dynamic = ["version"]

requires-python = ">=3.12"
dependencies = [
    "aiohttp ~= 3.13.3",
    "cachetools ~= 6.2.4",
    "colorlog ~= 6.10.1",
    "fastapi ~= 0.128.0",
    "langchain ~= 1.2.3",
    "langchain-community ~= 0.4.1",
    "langchain-core ~= 1.2.6",
    "langchain-openai ~= 1.1.7",
    "langfuse ~= 3.11.2",
    "langgraph ~= 1.0.5",
    "langsmith ~= 0.6.2",
    "fastmcp ~= 2.14.2",
    "pymilvus ~= 2.6.6",
    "python-dotenv ~= 1.2.1",
    "tiktoken ~= 0.12.0",
    "tqdm ~= 4.67.1",
    "uvicorn ~= 0.40.0",
]

[project.optional-dependencies]
dev = [
    "build ~= 1.4.0",
    "codespell ~= 2.4.1",
    "conventional-pre-commit ~= 4.3.0",
    "coverage ~= 7.13.1",
    "detect-secrets ~= 1.5.0",
    "git-cliff ~= 2.11.0",
    "pre-commit ~= 4.5.1",
    "pyright ~= 1.1.408",
    "ruff ~= 0.14.11",
]

[project.scripts]
wr-load = "wiki_rag.load.main:main"
wr-index = "wiki_rag.index.main:main"
wr-search = "wiki_rag.search.main:main"
wr-server = "wiki_rag.server.main:main"
wr-mcp = "wiki_rag.mcp_server.main:main"
wr-config-update = "wiki_rag.cli.config:config_update"

[project.urls]
Homepage = "https://github.com/moodlehq/wiki-rag"
Documentation = "https://github.com/moodlehq/wiki-rag"
Repository = "https://github.com/moodlehq/wiki-rag"

[tool.coverage.run]
branch = true
source = [
    "wiki_rag",
]
omit = [
    "tests"
]
command_line = "-m unittest discover -t . -s tests"

[tool.pyright]
include = [
    "wiki_rag"
]
exclude = [
    "tests"
]

[tool.ruff]
line-length = 88
indent-width = 4
preview = true

[tool.ruff.lint]
# https://docs.astral.sh/ruff/rules
select = [
    "E",    # pycodestyle (errors, all them)
    "F",    # pyflakes
    "I",    # imports
    "N",    # pep8 naming
    "D",    # pydocstyle
    "Q",    # flake8-quotes
    "ASYNC",# flake8-async
    "CPY",  # copyright header
    "EM",   # flake8-errmsg
    "UP",   # pyupgrade
]
ignore = [
    "D203", # 1 blank line required before class docstring (incompatible with D211, that we use)
    "D213", # Multi-line docstring summary should start at the second line (incompatible with D212, that we use)
]

[tool.ruff.lint.isort]
lines-between-types = 1

[tool.ruff.lint.pycodestyle]
max-line-length = 120

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401"] # Ignore the unused imports in init files because we use them to make easier package paths.
"tests/**/test_**.py" = ["D101", "D102", "D103"] # Don't require doc strings for tests

[tool.codespell]
ignore-words-list = [
  "upport" # Part of regular expressions in cliff.toml
]

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]

[tool.setuptools.packages.find]
include = ["wiki_rag"]
